lab03 Chebyshev done, functions redefined to work with float as well - file Chebyshev.cpp, in file Chebyshev_using_built_in_funcs.cpp code with built in functions

+++ lab03_Chebyshev/testChebyshev.cpp
@@ -3,12 +3,13 @@
 #include "Chebyshev.cpp"
 #include "consts.hpp"
 
+
 template<typename F>
 F checker() {
     F error;
     F max_error = 0;
     for (F x = -1; x <= 1; x += 0.05) {
-        size_t N = 20; // order = 20 is the best, because it gives the smallest error. The error is approximately equal to 10*eps
+        int N = 20; // seems like order = 20 is the best, because it gives the smallest error.
         error = std::abs(ADAAI::Chebyshev(x, N) - std::exp(x));
         if (error >= max_error) {
             max_error = error;
@@ -18,7 +19,7 @@ F checker() {
 }
 
 int main() {
-    std::cout << checker<float>() << '\n'; // gives max_error = 0. Maybe a problem with float TODO
+    std::cout << checker<float>() << '\n'; // for order = 20: max_error is approximately equal to 2*eps
 //    std::cout << ADAAI::c_Eps<float> << '\n';
     std::cout << checker<double>() << '\n'; // for order = 20: max_error is approximately equal to 10*eps
 //    std::cout << ADAAI::c_Eps<double> << '\n';
